feat: improve dired buffer updates
Mitigate a deficiency of `Dired' in that it relies upon `file-notify'
to trigger buffer updates. However, `file-notify' does not trigger
upon events that change the size of a file that is listed in a `Dired'
buffer. Thus, for example, when a tar archive is created, vanilla
`Dired' records its size as zero, and it will remain so until some
other action triggers a `revert-buffer' event."

This commit introduces defcustom diredc-update-interval.